LibWeb+LibWebView: Remove the old session history protocol
No callers remain for the WebContent-local queue or its traversal
messages. Remove them, keeping only the shared types and job entry
points used by UI-coordinated operations.
