+++ Libraries/LibWeb/CMakeLists.txt
@@ -728,7 +728,6 @@ set(SOURCES
     HTML/SelectItem.cpp
     HTML/SerializedPolicyContainer.cpp
     HTML/SessionHistoryEntry.cpp
-    HTML/SessionHistoryTraversalQueue.cpp
     HTML/SharedResourceRequest.cpp
     HTML/AnimatedBitmapDecodedImageData.cpp
     HTML/AnimatedDecodedImageData.cpp
