+++ Libraries/LibWeb/HTML/LocalTraversableNavigable.cpp
@@ -598,79 +598,6 @@ int LocalTraversableNavigable::get_the_used_step(int step) const
     return result.value();
 }
 
-// https://html.spec.whatwg.org/multipage/browsing-the-web.html#get-all-navigables-whose-current-session-history-entry-will-change-or-reload
-Vector<GC::Root<LocalNavigable>> LocalTraversableNavigable::get_all_local_navigables_whose_current_session_history_entry_will_change_or_reload(int target_step) const
-{
-    // 1. Let results be an empty list.
-    Vector<GC::Root<LocalNavigable>> results;
-
-    // 2. Let navigablesToCheck be « traversable ».
-    Vector<GC::Root<LocalNavigable>> navigables_to_check;
-    navigables_to_check.append(const_cast<LocalTraversableNavigable&>(*this));
-
-    // 3. For each navigable of navigablesToCheck:
-    while (!navigables_to_check.is_empty()) {
-        auto navigable = navigables_to_check.take_first();
-
-        // 1. Let targetEntry be the result of getting the target history entry given navigable and targetStep.
-        auto target_entry = navigable->get_the_target_history_entry_if_present(target_step);
-        if (!target_entry)
-            continue;
-
-        // 2. If targetEntry is not navigable's current session history entry or targetEntry's document state's reload
-        //    pending is true, then append navigable to results.
-        // AD-HOC: We don't want to choose a navigable that has ongoing traversal.
-        if ((target_entry != navigable->current_session_history_entry() || target_entry->document_state()->reload_pending()) && !navigable->ongoing_navigation().has<Traversal>()) {
-            results.append(*navigable);
-        }
-
-        // 3. If targetEntry's document is navigable's document, and targetEntry's document state's reload pending is
-        //    false, then extend navigablesToCheck with the child navigables of navigable.
-        if (target_entry->document_state()->document_id() == navigable->active_document_id() && !target_entry->document_state()->reload_pending()) {
-            navigables_to_check.extend(navigable->child_navigables());
-        }
-    }
-
-    // 4. Return results.
-    return results;
-}
-
-// https://html.spec.whatwg.org/multipage/browsing-the-web.html#getting-all-navigables-that-only-need-history-object-length/index-update
-Vector<GC::Root<LocalNavigable>> LocalTraversableNavigable::get_all_local_navigables_that_only_need_history_object_length_index_update(int target_step) const
-{
-    // NOTE: Other navigables might not be impacted by the traversal. For example, if the response is a 204, the currently active document will remain.
-    //       Additionally, going 'back' after a 204 will change the current session history entry, but the active session history entry will already be correct.
-
-    // 1. Let results be an empty list.
-    Vector<GC::Root<LocalNavigable>> results;
-
-    // 2. Let navigablesToCheck be « traversable ».
-    Vector<GC::Root<LocalNavigable>> navigables_to_check;
-    navigables_to_check.append(const_cast<LocalTraversableNavigable&>(*this));
-
-    // 3. For each navigable of navigablesToCheck:
-    while (!navigables_to_check.is_empty()) {
-        auto navigable = navigables_to_check.take_first();
-
-        // 1. Let targetEntry be the result of getting the target history entry given navigable and targetStep.
-        auto target_entry = navigable->get_the_target_history_entry_if_present(target_step);
-        if (!target_entry)
-            continue;
-
-        // 2. If targetEntry is navigable's current session history entry and targetEntry's document state's reload pending is false, then:
-        if (target_entry == navigable->current_session_history_entry() && !target_entry->document_state()->reload_pending()) {
-            // 1.  Append navigable to results.
-            results.append(navigable);
-
-            // 2. Extend navigablesToCheck with navigable's child navigables.
-            navigables_to_check.extend(navigable->child_navigables());
-        }
-    }
-
-    // 4. Return results.
-    return results;
-}
-
 // https://html.spec.whatwg.org/multipage/browsing-the-web.html#getting-all-navigables-that-might-experience-a-cross-document-traversal
 Vector<GC::Root<LocalNavigable>> LocalTraversableNavigable::get_all_local_navigables_that_might_experience_a_cross_document_traversal(int target_step) const
 {
