+++ Libraries/LibWebView/WebContentClient.cpp
@@ -1530,43 +1530,6 @@ void WebContentClient::did_request_traverse_the_history_by_delta(u64 page_id, i3
     (void)view->traverse_the_history_by_delta(delta, check_for_cancelation);
 }
 
-void WebContentClient::did_request_history_traversal_target_by_delta(u64 page_id, u64 request_id, i32 delta)
-{
-    Optional<i32> target_step;
-    if (auto* page_host = navigable_for_page(page_id)) {
-        auto target = page_host->top_level_traversable().session_history().traversal_target_for_delta(delta);
-        if (target.has_value())
-            target_step = target->target_step;
-    }
-
-    async_resolve_session_history_traversal_target(page_id, request_id, target_step);
-}
-
-void WebContentClient::did_request_traverse_the_history_to_step(u64 page_id, i32 step, Web::HistoryTraversalPrecheck history_traversal_precheck)
-{
-    auto* page_host = navigable_for_page(page_id);
-    if (!page_host)
-        return;
-
-    auto view = ViewImplementation::find_view_for_traversable(page_host->top_level_traversable());
-    if (!view.has_value())
-        return;
-
-    auto check_for_cancelation = history_traversal_precheck == Web::HistoryTraversalPrecheck::Needed
-        ? CheckForCancelation::Yes
-        : CheckForCancelation::IfWebContentCannotTraverseTarget;
-    (void)view->traverse_the_history_to_step(step, check_for_cancelation);
-}
-
-void WebContentClient::did_request_navigation_api_traversal_target(u64 page_id, u64 request_id, Web::HTML::CrossProcessId navigable_id, Utf16String navigation_api_key)
-{
-    Optional<i32> target_step;
-    if (auto navigable = hosted_navigable_for_page(page_id, navigable_id); navigable.has_value())
-        target_step = navigable->top_level_traversable().navigation_api_traversal_target(*navigable, navigation_api_key);
-
-    async_resolve_session_history_traversal_target(page_id, request_id, target_step);
-}
-
 void WebContentClient::did_request_webdriver_history_traversal(u64 page_id, u64 request_id, i32 delta)
 {
     if (auto view = view_for_page_id(page_id); view.has_value()) {
@@ -1996,18 +1959,6 @@ void WebContentClient::did_finalize_cross_document_navigation(u64 page_id, Web::
     navigable->top_level_traversable().finalize_cross_document_navigation(*navigable, move(history_entry), move(entry_to_replace_navigation_api_key));
 }
 
-void WebContentClient::did_set_current_session_history_step(u64 page_id, i32 current_session_history_step)
-{
-    auto* page_host = navigable_for_page(page_id);
-    if (!page_host)
-        return;
-
-    auto view = ViewImplementation::find_view_for_traversable(page_host->top_level_traversable());
-    if (!view.has_value())
-        return;
-    view->did_set_current_session_history_step({}, current_session_history_step);
-}
-
 Messages::WebContentClient::DidRequestUiProcessSessionHistoryForTestingResponse WebContentClient::did_request_ui_process_session_history_for_testing(u64 page_id)
 {
     if (auto view = view_for_page_id(page_id); view.has_value())
@@ -2035,18 +1986,6 @@ void WebContentClient::did_set_top_level_session_history(u64 page_id, bool accep
 {
     if (auto view = view_for_page_id(page_id); view.has_value())
         view->did_set_top_level_session_history({}, accepted, move(entries), move(used_steps), current_used_step_index);
-}
-
-void WebContentClient::did_traverse_the_history_to_step(u64 page_id, i32 step, bool step_was_available, Web::HTML::HistoryStepResult result)
-{
-    if (auto view = view_for_page_id(page_id); view.has_value())
-        view->did_traverse_the_history_to_step({}, step, step_was_available, result);
-}
-
-void WebContentClient::did_check_if_traverse_history_step_is_canceled(u64 page_id, u64 request_id, i32 step, Web::HTML::HistoryStepResult result)
-{
-    if (auto view = view_for_page_id(page_id); view.has_value())
-        view->did_check_if_traverse_history_step_is_canceled({}, request_id, step, result);
 }
 
 void WebContentClient::did_reset_session_history_for_testing(u64 page_id)
